requirements.txt
pdfplumber
pymupdf
pillow 